pyproject.toml
[project]
name = "rax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.0",
    "llama-index>=0.14.15",
    "llama-index-embeddings-openai>=0.3",
    "llama-index-readers-file>=0.4",
    "neo4j>=5.0",
    "openai>=2.21.0",
    "python-dotenv>=1.0",
    "uvicorn>=0.41.0",
]
